Package task-related classes into task
