+++ data/maga.txt
@@ -1,0 +1,1 @@
+E | 0 | hi | 2022-06-06
