+++ src/main/java/Ui.java
@@ -1,0 +1,21 @@
+import task.TaskList;
+
+import java.time.format.DateTimeParseException;
+import java.util.Scanner;
+
+public class Ui {
+    private final Scanner scanner = new Scanner(System.in);
+    private final Parser parser;
+
+    public Ui(TaskList taskList) {
+        parser =  new Parser(taskList, scanner);
+    }
+
+    public void start() {
+        String input = scanner.nextLine();
+        while (!input.equalsIgnoreCase("bye")) {
+            parser.handleInput(input);
+            input = scanner.nextLine();
+        }
+    }
+}
